refactor: 패키지구조를 수정하라
