+++ src/main/java/prgrms/project/stuti/domain/feed/repository/postcomment/PostCommentCustomRepository.java
@@ -1,4 +1,4 @@
-package prgrms.project.stuti.domain.feed.repository;
+package prgrms.project.stuti.domain.feed.repository.postcomment;
 
 import prgrms.project.stuti.domain.feed.service.dto.CommentParentContents;
 import prgrms.project.stuti.global.page.offset.PageResponse;
