+++ src/main/java/prgrms/project/stuti/domain/feed/repository/postcomment/PostCommentRepository.java
@@ -1,4 +1,4 @@
-package prgrms.project.stuti.domain.feed.repository;
+package prgrms.project.stuti.domain.feed.repository.postcomment;
 
 import org.springframework.data.jpa.repository.JpaRepository;
