--- src/main/java/prgrms/project/stuti/domain/feed/service/dto/PostLikeIdResponse.java
@@ -1,4 +1,0 @@
-package prgrms.project.stuti.domain.feed.service.dto;
-
-public record PostLikeIdResponse(Long postLikeId) {
-}
