+++ src/main/java/prgrms/project/stuti/domain/feed/service/response/PostLikeIdResponse.java
@@ -1,0 +1,4 @@
+package prgrms.project.stuti.domain.feed.service.response;
+
+public record PostLikeIdResponse(Long postLikeId) {
+}
