+++ src/main/java/prgrms/project/stuti/domain/feed/service/response/PostListResponse.java
@@ -1,4 +1,4 @@
-package prgrms.project.stuti.domain.feed.service.dto;
+package prgrms.project.stuti.domain.feed.service.response;
 
 import java.util.List;
